CLI argument parsing function
Zero-argument function to parse command-line arguments.  For this to be
used, the current usage for adding bot commands will need to be changed,
perhaps by using a Cog.  Returns the data as a dict.  All arguments are
marked as required, as opposed to checking for NoneType values after
parsing.  The argument `token` is positional, and the remainder are
keyword-only.